Core functional I/O threading primitives
Add core functional I/O threading primitives, in which all of the
core ManifoldCF I/O and exception handling logic is encapsulated. This
repetitive code is used all over the ManifoldCF code base, and if
this is accepted upstream, a lot of code can be simplified via the use
of these classes and lambdas.

For now, put this code in the O365 connector.