o365: additional meta-data, source and root paths
We obtain additional meta-data from the DriveItem and include it in
the RepositoryDocument. We also obtain additional site meta-data from
the site seed document via carry-down data, and include that as well
(which provides site name, display name, and URL).

Lastly we fix up and include the root and source paths.